fix(be): move rabbitmq module to libs
