+++ apps/backend/libs/rabbitmq/src/index.ts
@@ -1,0 +1,2 @@
+export * from './rabbitmq.module'
+export * from './mqtt.service'
