+++ apps/backend/tsconfig.json
@@ -30,7 +30,8 @@
       "@libs/logger": ["./libs/logger/src/index.ts"],
       "@libs/types": ["./libs/types/src/index.ts"],
       "@libs/instrumentation": ["./libs/instrumentation.ts"],
-      "@libs/storage": ["./libs/storage/src/index.ts"]
+      "@libs/storage": ["./libs/storage/src/index.ts"],
+      "@libs/rabbitmq": ["./libs/rabbitmq/src/index.ts"]
     }
   }
 }
